refacto(Input handling): reorganize files according to project's file organization philosophy, remove redundant functions, make AHandler class even more generic with template type use
