+++ src/Input/Keyboard.hpp
@@ -7,35 +7,28 @@
 
 #pragma once
 
-#include "Handler.hpp"
+#include "../../includes/Input/AHandler.hpp"
 
 namespace input {
-    class KeyboardHandler : public AHandler {
+    class KeyboardHandler : public AHandler<SDL_Keycode> {
         public:
             KeyboardHandler() : AHandler(Type::KEYBOARD) {
-                keyboardBindings = {
-                    {SDLK_UP, Generic::UP},
-                    {SDLK_DOWN, Generic::DOWN},
-                    {SDLK_LEFT, Generic::LEFT},
-                    {SDLK_RIGHT, Generic::RIGHT},
+                inputBindings = {
+                    {SDLK_w, Generic::UP},
+                    {SDLK_s, Generic::DOWN},
+                    {SDLK_a, Generic::LEFT},
+                    {SDLK_d, Generic::RIGHT},
                     {SDLK_e, Generic::INTERACT},
                     {SDLK_SPACE, Generic::ATTACK},
-                    {SDLK_i, Generic::INVENTORY},
+                    {SDLK_TAB, Generic::INVENTORY},
                     {SDLK_RETURN, Generic::SELECT},
                     {SDLK_ESCAPE, Generic::PAUSE}
                 };
             };
+            ~KeyboardHandler() = default;
 
         private:
-            std::unordered_map<SDL_Keycode, Generic> keyboardBindings;
-
-            Generic getGenericFromEvent(const SDL_Event &event) const {
-                auto it = keyboardBindings.find(event.key.keysym.sym);
-                return (it != keyboardBindings.end()) ? it->second : Generic::VOID;
-            }
-
-            State getState(const SDL_Event &event) const {
-                return (event.type == SDL_KEYDOWN) ? State::PRESSED : State::RELEASED;
-            }
+            Generic getGenericFromEvent(const SDL_Event &event) const;
+            State getState(const SDL_Event &event) const;
     };
 }
